Changed test names
removed tmp file from git
